// src/modules/delivery-partner/types/delivery-partner.types.ts
export enum DeliveryPartnerErrorCode {
  DELIVERY_PARTNER_NOT_FOUND = 'DELIVERY_PARTNER_NOT_FOUND',
  DELIVERY_PARTNER_SUSPENDED = 'DELIVERY_PARTNER_SUSPENDED',
  DELIVERY_PARTNER_INACTIVE = 'DELIVERY_PARTNER_INACTIVE',
  SERVICE_AREA_NOT_FOUND = 'SERVICE_AREA_NOT_FOUND',
  SERVICE_AREA_ALREADY_EXISTS = 'SERVICE_AREA_ALREADY_EXISTS',
  INVALID_SERVICE_AREA = 'INVALID_SERVICE_AREA',
  AVAILABILITY_NOT_FOUND = 'AVAILABILITY_NOT_FOUND',
  AVAILABILITY_CONFLICT = 'AVAILABILITY_CONFLICT',
  INVALID_AVAILABILITY = 'INVALID_AVAILABILITY',
  DOCUMENT_NOT_FOUND = 'DOCUMENT_NOT_FOUND',
  DOCUMENT_INVALID = 'DOCUMENT_INVALID',
  DOCUMENT_STATE_INVALID = 'DOCUMENT_STATE_INVALID',
  DELIVERY_PARTNER_ACCESS_DENIED = 'DELIVERY_PARTNER_ACCESS_DENIED',
}

export enum DeliveryPartnerAuditEventType {
  DELIVERY_PARTNER_PROFILE_UPDATED = 'DELIVERY_PARTNER_PROFILE_UPDATED',
  DELIVERY_SERVICE_AREA_CREATED = 'DELIVERY_SERVICE_AREA_CREATED',
  DELIVERY_SERVICE_AREA_UPDATED = 'DELIVERY_SERVICE_AREA_UPDATED',
  DELIVERY_SERVICE_AREA_REMOVED = 'DELIVERY_SERVICE_AREA_REMOVED',
  DELIVERY_AVAILABILITY_CREATED = 'DELIVERY_AVAILABILITY_CREATED',
  DELIVERY_AVAILABILITY_UPDATED = 'DELIVERY_AVAILABILITY_UPDATED',
  DELIVERY_AVAILABILITY_REMOVED = 'DELIVERY_AVAILABILITY_REMOVED',
  DELIVERY_DOCUMENT_SUBMITTED = 'DELIVERY_DOCUMENT_SUBMITTED',
  DELIVERY_DOCUMENT_REMOVED = 'DELIVERY_DOCUMENT_REMOVED',
}